refactor: move tests and make PR changes
